backend/requirements.txt
asgiref==3.9.0
black==25.1.0
Brotli==1.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
Django==5.2.4
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-grappelli==4.0.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.4.0
drf-nested-routers==0.94.2
gunicorn==23.0.0
idna==3.10
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pip==25.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
PyJWT==2.10.1
requests==2.32.4
setuptools==78.1.1
sqlparse==0.5.3
urllib3==2.5.0
wheel==0.45.1
whitenoise==6.9.0
